bluetooth: a2dp: Remove "congestion" logic
I believe that this congestion logic was added to try to put
a bandaid on the bug that was fixed in this commit:

Author: Christopher R. Palmer <crpalmer@gmail.com>
Date:   Wed Feb 25 15:27:29 2015 -0500

    bluetooth: a2dp: Don't count generated data when we can't send it

Since it was incorrectly accumulating more data then we would
try to hammer the sink device.  However, after fixing that bug,
the total volume of data to send should now be sane and we don't
have to worry about overwhelming downstream devices.

Change-Id: I3b30ed0744645345cde29fe32d2bc0ff6ef043d3